clean up repeat code
